Add files via upload
The main objective of this project is to use an ESP32 low-cost System on Chip (SoC) Microcontroller and a MAX30102 sensor to measure the Heart Rate and Body temperature of an individual and upload the readings to a database (in this case we used Firebase). The data can then be further processed for other purposes as demonstrated in various tutorials online.